Remove duplicate Unauthorized component
