+++ frontend/src/components/NotFound.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const NotFound = (): JSX.Element => (
+  <img
+    src="/images/404.png"
+    style={{ width: "100%", height: "auto" }}
+    alt="404 Not Found"
+  />
+);
+
+export default NotFound;
