+++ frontend/src/components/index.tsx
@@ -2,4 +2,5 @@ export { default as Navigation } from "./Navigation";
 export { default as Footer } from "./Footer";
 export { default as Countdown } from "./Countdown";
 export { default as Block } from './Block';
-export { default as Unauthorized } from './Unauthorized'
+export { default as Unauthorized } from './Unauthorized';
+export { default as NotFound } from './NotFound';
