+++ frontend/src/pages/gold/Connect.tsx
@@ -5,7 +5,7 @@ import { Block } from "../../components"
 const Connect = (): JSX.Element => (
   <Block size='xs-12'>
     <p>You have not yet attached your Scratch® account to your profile. Please enter it below.</p>
-    <Form action="/gold/connect" method="POST" class="ui action input">
+    <Form action="/gold/connect" method="POST" className="ui action input">
       <Input placeholder="Scratch Username" />
       <Button content='Connect' floated='right' type="submit" disabled />
     </Form>
