+++ frontend/src/pages/index.tsx
@@ -4,11 +4,11 @@ import Home from "./Home";
 import About from "./About";
 import Login from "./Login";
 import Help from "./Help";
-import NotFound from "./NotFound";
 import DefaultNavigation from "./DefaultNavigation";
+import Clarifications from './Clarifications';
 import { Container } from "semantic-ui-react";
-
 import { Route, Switch } from "react-router-dom";
+import { NotFound } from "../components";
 
 const Index = (): JSX.Element => (
   <>
@@ -20,6 +20,7 @@ const Index = (): JSX.Element => (
         <Route path="/about" component={About} />
         <Route path="/login" component={Login} />
         <Route path="/help" component={Help} />
+        <Route path='/clarifications' component={Clarifications} />
         <Route default component={NotFound} />
       </Switch>
     </Container>
